+++ src/utils/jbig2/arithmetic_decoder.js
