+++ src/utils/jbig2/ccitt.js
