+++ src/utils/jbig2/compatibility.js
