+++ src/utils/jbig2/core_utils.js
