+++ src/utils/jbig2/is_node.js
