+++ src/utils/jbig2/jbig2.js
