+++ src/utils/jbig2/jbig2_stream.js
