+++ src/utils/jbig2/primitives.js
