+++ src/utils/jbig2/stream.js
